requirements.txt
aiogram==2.8
aiohttp==3.7.4
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.6
pycodestyle==2.7.0
python-dotenv==0.17.1
pytz==2020.1
typing-extensions==3.10.0.0
yarl==1.4.2
